content/canvas/test/test_2d.type.delete.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.type.delete</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

delete window.CanvasRenderingContext2D;
ok(window.CanvasRenderingContext2D !== undefined, "window.CanvasRenderingContext2D !== undefined");

if (!_deferred) SimpleTest.finish();
});
</script>
